+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Observations
 * Al apretar el botón de cerrar en el modal de cobros, no se cierra.
 * Hay inconsistencias entre el frontend de cobros y el backend. Se calcula a mano el monto restante y el total abonado cuando esta información ya es devuelta por el backend.
+* El manejo de inputs como tipo "number" en el modal `CrearEditarCobroModal.tsx` da un mal UX al no permitir borrar el 0 inicial, produciendo situaciones donde el usuario quiere escribir "1" pero como el 0 no se puede borrar se escribe "01".
+* El modal `CrearEditarCobroModal.tsx` no tiene documentación.
 
 ### Changed
 * El callback `closeCobrosModal` en `GestionPedidosPage.tsx` afectaba al estado equivocado. Se corrigió para que afecte a `setIsCobrosModalOpen`.
@@ -11,13 +13,16 @@
 ### Refactored
 * Ahora en `GestionCobroModal.tsx` se hace provecho de la información del backend para calcular `montoRestante` y `totalAbonado` en vez de calcularlos a mano.
 * Ajusta el tipo `Cobro` en `types/models.ts`
+* Cambia el tipo de los inputs del formulario `CrearEditarCobroModal.tsx` de "number" a "string"
+
+### Added
+* Documentación estilo doxygen a `CrearEditarCobroModal.tsx`.
 
 ### Affects
 * `GestionPedidosPage.tsx`
 * `GestionCobrosModal.tsx`
 * `models.ts`
-
-###
+* `CrearEditarCobroModal.tsx`
 
 ## [ fix/merge/cobros-ui ] - 2025-10-13